requirements.txt
python-telegram-bot==20.8
openai==1.3.5
praw==7.6.0
requests==2.32.3
python-dotenv==1.0.1
